+++ tests/apps/search/test_viewsets_courses.py
@@ -290,6 +290,7 @@ class CoursesViewsetsTestCase(CMSTestCase):
             _source=[
                 "absolute_url",
                 "categories",
+                "code",
                 "cover_image",
                 "duration",
                 "effort",